Weapon Statistics & Particles
In this commit, I have created 3 prefabs of unique guns (Shotgun, SMG & LMG) which is provided with a script of the weapon base with suitable data such as the weapon stat, particle system and raycast.

In each prefab has 3 child objects:
- Particle System: Create particles that represents the bullet speed and spread.
- Raycast: Which has nothing in currently. however is placed in the weapon based script under raycast
- Model: Which has nothing in currently.

In the particles I have edited:
- Emission (Shotgun): This shoots each particle at the same time rather than individually
- Collision: This when colliding with an object will destroy the particle.

The rest of the edit involves changing the speed stimulation, speed range and maximum particles.