Issue Identified & Possibly Solved
There was an issue with the prefab towards both scripts (WeaponBase & WeaponStats) from the prefab.

I couldn't figure out a solution towards the cause of this problem, however I did managed to possible solved this by returning the scripts back inside the prefabs.

Now with this commit should have the weapon base script inside each of the prefabs along with the suitable data in each variable along with the weapon stats.