Strange Issue with Weapon Prefabs
So this issue is odd, whenever I close unity onto one device and then open up again the prefabs are fine.

However when using another device and loading up unity the scripts inside the prefabs and the weapon stat data goes missing.

I have no clue on the issue, however I return their values again. If this issue occurs for other users then it has to be due to the either the weaponbase or weaponstat script.