Able to Read Primary & Secondary Abilities
This commit, I managed to ensure that whenever the player presses either 1 or 2 (Primary/Secondary button) a public function would be able to call to able to read both primary and secondary abilities.

This tasks is not 100% complete from my end as there may need to be more instructions inside these functions.