Providing Public Events
In this commit I have updated the primary and secondary to have events, which will invoke the events one the function is performed by pressing either of the abilities.

A temporary test script called "TestEventFire" in my TestScene was use to check if the events worked. The public functions used were "FirePrimary" & "FireSecondary"

Each event provided the test script with their suitable function to call when the event has been invoked.