Improve on the playerModel and playerData Scripts
In this commit, I updated the scripts based on providing a method to increase the blood and ensure the blood stays as the maximum value (This was inside of the playerModel).

I also included Iron Resources, which is similar to the blood functions except "iron" (Also inside of the playerModel)

And finally I have added a starting iron stat and max iron resource to have the player start with certain amount of iron and the amount the player can carry.

Testing hasn't been fully completed yet however so it's not defined as done. However testing toward the get values are looking positive.

Next commit will be to implement my test script for the resources to check if the resources increase/decrease and check if their values don't go above the max resource value or below 0.