Passive Ability (Controller, Base & Ability Script)
In this commit I have created a folder called "passiveAbilities" in this folder includes 3 scripts:
passiveController - has voids that are called based on the input of one of the four magic buttons
passiveBase - has a virtual void that can be inherit from the passiveAbility Class
passiveAbility - has the inheritance of the passiveBase Class.

There isn't any implementations inside of the voids currently, which will be implemented soon.