Water Colliding
In this commit, I manage to provide a solution towards the player to collide with the water once the current magic is nature. This was done by testing if a log comes up when entering the water with the nature ability.

Although this does help the nature passive ability, my tasks is nowhere complete yet due to technical issues and my own struggle towards finding a suitable solution.