Nature Passive Ability
In this commit, I have manage to provide a method that uses the sphere collider from the child to detect if the radius has entered or exit any water.

If player has entered then this will increase the max health and reduce the damage.

Otherwise if player has exited then this will return the values back to normal.

I have fixed an issue with the MaxHealth since this didn't reference the maxHealth variable.