Interact Event, Particle & Height & Spawn Manager
The interact event is whenever the player has enter area of the interact event and presses run whilst inside (There isn't a input action for interact) enemies will spawn.

Particle is included in the SpawnData scriptable object, this is then added into the Spawn method in the Spawn Manager script which if the particle has a particle system then the particle will play.

The height is a float which is included in the SpawnData scriptable object and  has a Vector3 method inside of the Spawn Manager in which the y axis of the transform position is equal to the height float in the SpawnData, this allows the flying enemies to spawn above the ground compare to the swarmer and tank enemies.

The Spawn manager has been tested with the Spawn Data for each event having test prefabs I have made:
- Circle = Swarmer
- Box = Tank
- Cylinder = Flying

All three events worked from their criteria to spawn enemies and the spawn manager was capable of spawning suitable amount of enemies and type of enemies included in the Spawn Data.

A test event script was made to check if when the TimedEvent script starts its' coroutine it will spawn the enemies if the player didn't react.